requirements.txt
requests
beautifulsoup4
gspread==6.0.2
google-auth==2.29.0
python-dotenv